Create/Edit/Delete user action should be validated for the admin only

+++ src/lib/action.ts
@@ -2,9 +2,20 @@
 
 import { UserForm } from "./types";
 import { adminDb } from "./firebase_admin";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
+import { NextResponse } from "next/server";
+import { STATUS_CODES } from "http";
 
 export async function deleteUser(id: string) {
+  const session = await getServerSession(authOptions);
+  const user = session?.user;
+  console.log(user);
+  if (user?.role == "user") {
+    return NextResponse.json({ msg: "msg", code: 403 });
+  }
   console.log("id", id);
+
   try {
     await adminDb.collection("users").doc(id).delete();
 
@@ -16,10 +27,14 @@ export async function deleteUser(id: string) {
   }
 }
 
-export async function editUser(
-  id: string,
-  formData: UserForm
-): Promise<boolean> {
+export async function editUser(id: string, formData: UserForm) {
+  const session = await getServerSession(authOptions);
+  const user = session?.user;
+  if (user?.role !== "admin") {
+    // If user is not logged in or not an admin, return msg response
+    return NextResponse.json({ msg: "msg", code: 403 });
+  }
+
   console.log("id", id);
   console.log("formData", formData);
   try {
@@ -31,7 +46,13 @@ export async function editUser(
     throw new Error("Failed to update user");
   }
 }
-export async function createUser(formData: UserForm): Promise<boolean> {
+export async function createUser(formData: UserForm) {
+  const session = await getServerSession(authOptions);
+  const user = session?.user;
+  if (!user || user.role !== "admin") {
+    return NextResponse.json({ msg: "msg", code: 403 });
+  }
+
   console.log("formData", formData);
   try {
     await adminDb.collection("users").add(formData);
